🧹 Fold the superadmin refusal into a shared concern
The blocked flag was only ever set to true, so a Site that refused one
assignment kept failing validation even after a later, valid one. It now
records the outcome of each assignment, including the permitted case.

Site and Account carried identical copies of the validation, so both now
include SuperadminRemovalGuard instead. That also makes the validation
private on Account, which it wasn't.
