+++ app/models/account.rb
@@ -6,6 +6,7 @@ class Account < ApplicationRecord
   include AccountEndpoints
   include AccountSettings
   include AccountCname
+  include SuperadminRemovalGuard
   attr_readonly :tenant, :public_demo_tenant
 
   has_many :sites, dependent: :destroy
@@ -51,7 +52,6 @@ class Account < ApplicationRecord
 
   # NEW: Validate that search-only accounts have at least one full account selected
   validate :search_only_must_have_full_accounts, if: :search_only?
-  validate :superadmin_removal_permitted
 
   after_save :schedule_jobs_if_settings_changed
 
@@ -182,14 +182,8 @@ class Account < ApplicationRecord
       site.superadmin_emails = emails
       # Carry the refusal up so the proprietor form, which edits an Account,
       # fails validation rather than silently reporting success.
-      @superadmin_removal_blocked = site.superadmin_removal_blocked?
+      self.superadmin_removal_blocked = site.superadmin_removal_blocked?
     end
-  end
-
-  def superadmin_removal_permitted
-    return unless @superadmin_removal_blocked
-
-    errors.add(:superadmin_emails, :cannot_remove_last_superadmin)
   end
 
   def cache_api?
