+++ app/models/site.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
 
 class Site < ApplicationRecord
+  include SuperadminRemovalGuard
+
   resourcify
 
   validates :application_name, presence: true, allow_nil: true
-  validate :superadmin_removal_permitted
 
   # Allow for uploading of site's banner image
   mount_uploader :banner_image, Hyku::AvatarUploader
@@ -55,13 +56,6 @@ class Site < ApplicationRecord
     User.with_role(:superadmin, self).pluck(:email)
   end
 
-  # True when the last assignment declined to remove a superadmin because it
-  # would have left a public demo tenant with none. Account reads this to
-  # raise the same error on itself, since the proprietor form edits an Account.
-  def superadmin_removal_blocked?
-    @superadmin_removal_blocked.present?
-  end
-
   # Update superadmin emails associated with this site
   # @param [Array<String>] Array of user emails
   def superadmin_emails=(emails)
@@ -69,10 +63,8 @@ class Site < ApplicationRecord
     existing_superadmin_emails = superadmin_emails
     new_superadmin_emails = emails - existing_superadmin_emails
     removed_superadmin_emails = existing_superadmin_emails - emails
-    if removed_superadmin_emails.present? && strips_last_superadmin?(emails)
-      @superadmin_removal_blocked = true
-      removed_superadmin_emails = []
-    end
+    self.superadmin_removal_blocked = removed_superadmin_emails.present? && strips_last_superadmin?(emails)
+    removed_superadmin_emails = [] if superadmin_removal_blocked?
     add_superadmins_by_email(new_superadmin_emails) if new_superadmin_emails.present?
     remove_superadmins_by_email(removed_superadmin_emails) if removed_superadmin_emails.present?
   end
@@ -126,16 +118,8 @@ class Site < ApplicationRecord
     end
   end
 
-  def superadmin_removal_permitted
-    return unless @superadmin_removal_blocked
-
-    errors.add(:superadmin_emails, :cannot_remove_last_superadmin)
-  end
-
   # Public demo tenants must always keep at least one site-scoped superadmin.
-  # Role changes apply on assignment rather than on save, so the removal is
-  # skipped rather than performed and undone; the validation above then fails
-  # the save.
+  # The refusal itself is carried by SuperadminRemovalGuard.
   # @param [Array<String>] requested_emails the full email list being assigned
   def strips_last_superadmin?(requested_emails)
     return false unless account&.public_demo_tenant?
